requirements_render.txt
python-telegram-bot==21.4
flask==3.0.3
requests==2.31.0
gunicorn==21.2.0
